pyproject.toml
[project]
name = "orche"
version = "0.3.0"
description = "A Python orchestrator for Docker Compose stacks"
authors = [
    {name = "Pietro Agazzi", email = "pietro.agazzi@prconsulting.eu"}
]
readme = "README.md"
requires-python = ">=3.10,<4"
dependencies = [
    "python-on-whales>=0.80.0,<0.81.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0",
    "GitPython>=3.1.0",
    "click>=8.1.0",
]

[project.scripts]
orche = "orche.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-timeout>=2.3.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
    "types-PyYAML>=6.0",
    "types-pygments>=2.19.0",
    "pre-commit>=4.0.0",
]

[tool.poetry]
packages = [{include = "orche"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --timeout=300"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = []

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
files = ["orche", "tests"]
